requirements.txt
# Core
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Data & ML
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.3
joblib==1.3.2

# API & Web
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
httpx==0.25.2

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
redis==5.0.1

# MLOps
mlflow==2.9.2
prefect==2.14.10

# Monitoring
prometheus-client==0.19.0
evidently==0.4.11

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Utilities
python-multipart==0.0.6
